move d3 logic into own class
